pyproject.toml
[project]
name = "corporate-manager-fixed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler==3.11.2",
    "dlib>=20.0.0",
    "ephem==4.2",
    "face-recognition==1.3.0",
    "face-recognition-models",
    "flask==3.1.2",
    "flask-limiter>=4.1.1",
    "flask-wtf>=1.2.2",
    "hijri-converter==2.3.2.post1",
    "hijridate==2.3.0",
    "holidays==0.91",
    "mysql-connector-python==9.5.0",
    "numpy==2.2.6",
    "opencv-python==4.12.0.88",
    "pillow==12.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.2.2",
    "pytz==2025.2",
    "tzdata==2025.3",
    "tzlocal==5.3.1",
]

[tool.uv.sources]
face-recognition-models = { git = "https://github.com/ageitgey/face_recognition_models" }
